Move TrustDetails to TrustRepository
