+++ DfE.FindInformationAcademiesTrusts.Data.AcademiesDb/Repositories/TrustRepository.cs
@@ -1,4 +1,5 @@
 using DfE.FindInformationAcademiesTrusts.Data.AcademiesDb.Contexts;
+using DfE.FindInformationAcademiesTrusts.Data.AcademiesDb.Extensions;
 using DfE.FindInformationAcademiesTrusts.Data.RepositoryDto;
 using Microsoft.EntityFrameworkCore;
 
@@ -7,6 +8,7 @@ namespace DfE.FindInformationAcademiesTrusts.Data.AcademiesDb.Repositories;
 public interface ITrustRepository
 {
     Task<TrustSummaryRepoDto?> GetTrustSummaryAsync(string uid);
+    Task<TrustDetailsRepoDto> GetTrustDetailsAsync(string uid);
 }
 
 public class TrustRepository(IAcademiesDbContext academiesDbContext) : ITrustRepository
@@ -19,5 +21,54 @@ public class TrustRepository(IAcademiesDbContext academiesDbContext) : ITrustRep
             .SingleOrDefaultAsync();
 
         return details is null ? null : new TrustSummaryRepoDto(details.Name, details.Type);
+    }
+
+    public async Task<TrustDetailsRepoDto> GetTrustDetailsAsync(string uid)
+    {
+        var giasGroup = await academiesDbContext.Groups
+            .Where(g => g.GroupUid == uid)
+            .Select(giasGroup => new
+            {
+                giasGroup.GroupUid,
+                giasGroup.GroupId,
+                giasGroup.Ukprn,
+                giasGroup.CompaniesHouseNumber,
+                giasGroup.GroupType,
+                giasGroup.GroupContactStreet,
+                giasGroup.GroupContactLocality,
+                giasGroup.GroupContactTown,
+                giasGroup.GroupContactPostcode,
+                giasGroup.IncorporatedOnOpenDate
+            })
+            .SingleAsync();
+
+        var regionAndTerritory = await GetRegionAndTerritoryAsync(uid);
+
+        var trustDetailsDto = new TrustDetailsRepoDto(
+            giasGroup.GroupUid!, //Searched by this field so it must be present
+            giasGroup.GroupId,
+            giasGroup.Ukprn,
+            giasGroup.CompaniesHouseNumber,
+            giasGroup.GroupType!, //Enforced by EF filter
+            string.Join(", ", new[]
+            {
+                giasGroup.GroupContactStreet,
+                giasGroup.GroupContactLocality,
+                giasGroup.GroupContactTown,
+                giasGroup.GroupContactPostcode
+            }.Where(s => !string.IsNullOrWhiteSpace(s))),
+            regionAndTerritory,
+            giasGroup.IncorporatedOnOpenDate.ParseAsNullableDate()
+        );
+
+        return trustDetailsDto;
+    }
+
+    private async Task<string> GetRegionAndTerritoryAsync(string uid)
+    {
+        return await academiesDbContext.MstrTrusts
+            .Where(m => m.GroupUid == uid)
+            .Select(m => m.GORregion)
+            .SingleOrDefaultAsync() ?? string.Empty;
     }
 }
